Fancy header (#16)
* Enhance pageHeader component styling

Add new styles for the page header, including branding and navigation elements. Implement gradients, hover effects, and layout adjustments to improve the visual presentation and user experience of the header section.

* Add 'Blog' link to navigation and adjust media query breakpoint

In the pageHeader component, a new 'Blog' link has been added to the main navigation menu. Additionally, the media query breakpoint for larger screens has been updated from 600px to 670px to improve layout responsiveness.

* Update pageHeader component styles and add logo text overlay

In the pageHeader component, the logo font size has been increased to 3rem, and new styles have been added for a text overlay that displays "mjk @ web" with a gradient effect. The logo's position and display properties have also been adjusted for improved visual presentation.

* Update pageHeader styles with margin adjustments and border-radius enhancements

In the pageHeader component, the margin-bottom has been set to -10px for better spacing, and the border-radius for navigation items has been updated to 15px. Additionally, hover effects for the first and last navigation items have been added to create a more visually appealing design.

* Update pageHeader styles with new gradient and shadow effects

In the pageHeader component, the background gradient for the main navigation has been updated for a fresher look, and enhancements have been made to the box-shadow properties for improved depth and visual appeal. These changes aim to elevate the overall design of the header section.